docs(community): process files separately to data
